requirement.txt
tensorflow
tensorflow-addons
opencv-python
scikit-learn
joblib
numpy
pillow
vit_keras
hydra-core
annoy
